Correction

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	long long int sum = 0, num1 = 0, num2 = 1, total = 0;
+	long int sum = 0, num1 = 0, num2 = 1, total = 0;
 
 	while (1)
 	{
@@ -20,6 +20,6 @@ int main(void)
 		num1 = num2;
 		num2 = sum;
 	}
-	printf("%lld\n", total);
+	printf("%ld\n", total);
 	return (0);
 }
